Revert "Remove redundant conversions"
oops, forgot that ssize_t and size_t are platform-dependent.

This reverts commit eefd23651aa21b22a1cf3eb77da4010a35801066.

+++ ktor-io/mingwX64/src/CharsetMingw.kt
@@ -85,8 +85,8 @@ internal actual fun CharsetEncoder.encodeImpl(input: CharSequence, fromIndex: In
                              * outbytesleft is decremented by the number of bytes written
                              */
                             val convertResult =
-                                iconv(cd, inbuf.ptr, inbytesleft.ptr, outbuf.ptr, outbytesleft.ptr)
-                            if (convertResult == MAX_SIZE) {
+                                iconv(cd, inbuf.ptr, inbytesleft.ptr, outbuf.ptr, outbytesleft.ptr).toULong()
+                            if (convertResult == MAX_SIZE.toULong()) {
                                 checkIconvResult(posix_errno())
                             }
 
@@ -132,7 +132,7 @@ public actual fun CharsetDecoder.decode(input: Source, dst: Appendable, max: Int
                             outbytesleft.value = (chars.size * 2).toULong()
 
                             val result = iconv(cd, inbuf.ptr, inbytesleft.ptr, outbuf.ptr, outbytesleft.ptr)
-                            if (result == MAX_SIZE) {
+                            if (result == MAX_SIZE.toULong()) {
                                 checkIconvResult(posix_errno())
                             }
